+++ src/UserProfileManagement/UserProfile.Application/UserProfiles/Commands/DeleteUserProfileCommand.cs
@@ -1,20 +1,35 @@
-﻿using System.Threading;
+﻿using System;
+using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
+using BuildingBlocks.Core.Interfaces;
+using Hive.Common.Core;
 using Hive.Common.Core.Exceptions;
+using Hive.Common.Core.Interfaces;
+using Hive.Common.Core.Security.Handlers;
 using Hive.UserProfile.Application.Interfaces;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.Extensions.Logging;
 
 namespace Hive.UserProfile.Application.UserProfiles.Commands
 {
     public record DeleteUserProfileCommand(int UserProfileId) : IRequest;
 
-    public class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfileCommand>
+    public class DeleteUserProfileCommandHandler : AuthorizationRequestHandler<Domain.Entities.UserProfile>,  IRequestHandler<DeleteUserProfileCommand>
     {
         private readonly IUserProfileDbContext _dbContext;
+        private readonly IFileService _fileService;
+        private readonly ILogger<DeleteUserProfileCommandHandler> _logger;
 
-        public DeleteUserProfileCommandHandler(IUserProfileDbContext dbContext)
+        public DeleteUserProfileCommandHandler(IUserProfileDbContext dbContext,
+            ICurrentUserService currentUserService, IAuthorizationService authorizationService,
+            IFileService fileService,
+            ILogger<DeleteUserProfileCommandHandler> logger) : base(currentUserService, authorizationService)
         {
-            _dbContext = dbContext;
+            _dbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
+            _fileService = fileService ?? throw new ArgumentNullException(nameof(fileService));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }
         
         public async Task<Unit> Handle(DeleteUserProfileCommand request, CancellationToken cancellationToken)
@@ -23,12 +38,27 @@ namespace Hive.UserProfile.Application.UserProfiles.Commands
 
             if (userProfile is null)
             {
-                throw new NotFoundException(nameof(Domain.UserProfile), request.UserProfileId);
+                _logger.LogWarning("User Profile with id: {@Id} was not found.", request.UserProfileId);
+                throw new NotFoundException(nameof(Domain.Entities.UserProfile), request.UserProfileId);
+            }
+            
+            var result = await base.AuthorizeAsync(userProfile,  new [] {"OnlyOwnerPolicy"});
+            
+            if (!result.All(s => s.Succeeded))
+            {
+                throw new ForbiddenAccessException();
+            }
+
+            if (!string.IsNullOrEmpty(userProfile.AvatarUri))
+            {
+                var deleted = await _fileService.DeleteAsync("user-avatars", userProfile.AvatarUri, cancellationToken);                
             }
 
             _dbContext.UserProfiles.Remove(userProfile);
             await _dbContext.SaveChangesAsync(cancellationToken);
             
+            _logger.LogWarning("User Profile with id: {@Id} was deleted.", request.UserProfileId);
+
             return Unit.Value;
         }
     }
